[MOD] Properly Fix width of menu & implement hierarchy items

+++ Editor/GUI/Drawables/IOrderedDrawable.cs
@@ -6,7 +6,7 @@ namespace Rhinox.GUIUtils.Editor
 {
     public interface IOrderedDrawable
     {
-        int Order { get; set; }
+        float Order { get; set; }
         ICollection<TAttribute> GetMemberAttributes<TAttribute>() where TAttribute : Attribute;
         void Draw();
         void Draw(Rect rect);
